+++ Container.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import View from './View.js';
-import TranslationWordsFetchData from './FetchData';
-import ScripturePaneFetchData from '../ScripturePane/FetchData';
-import VerseCheckFetchData from '../VerseCheck/FetchData';
-import TranslationHelpsFetchData from '../TranslationHelps/FetchData';
+// import TranslationWordsFetchData from './FetchData';
+// import ScripturePaneFetchData from '../ScripturePane/FetchData';
+// import VerseCheckFetchData from '../VerseCheck/FetchData';
+// import TranslationHelpsFetchData from '../TranslationHelps/FetchData';
+import FetchData from './FetchData'
 // Api Consts
 const api = window.ModuleApi;
 const NAMESPACE = "ImportantWords";
@@ -24,16 +25,13 @@ class Container extends React.Component {
 
   componentWillMount() {
     let { resourcesReducer, projectDetailsReducer, actions } = this.props;
-    let { progress, addNewBible, addNewResource, setModuleSettings } = actions;
+    let { progress, addNewBible, addNewResource, setModuleSettings, addGroupData, addGroupIndex } = actions;
     let props = {
       params: projectDetailsReducer.params,
       manifest: projectDetailsReducer.manifest,
       bibles: resourcesReducer.bibles
     };
-    ScripturePaneFetchData(addNewBible, addNewResource, props, progress, setModuleSettings);
-    TranslationHelpsFetchData(progress);
-    VerseCheckFetchData(progress);
-    TranslationWordsFetchData(addNewBible, addNewResource, props, progress);
+    FetchData(this.props);
   }
 
   saveProjectAndTimestamp() {
@@ -169,50 +167,6 @@ class Container extends React.Component {
     return (
       <div></div>
     );
-  }
-}
-
-
-
-/**
-* @description Compares two string alphabetically.
-* @param {string} first - string to be compared against.
-* @param {string} second - string to be compared with.
-*/
-function stringCompare(first, second) {
-  if (first < second) {
-    return -1;
-  } else if (first > second) {
-    return 1;
-  } else {
-    return 0;
-  }
-}
-
-/**
-* @description - Binary search of the list. I couldn't find this in the native methods of an array so
-* I wrote it
-* @param {array} list - array of items to be searched
-* @param {function} boolFunction - returns # < 0, # > 0. or 0 depending on which path the
-* search should take
-* @param {int} first - beginnging of the current partition of the list
-* @param {int} second - end of the current partition of the list
-*/
-function search(list, boolFunction, first = 0, last = -1) {
-  if (last == -1) {
-    last = list.length;
-  }
-  if (first > last) {
-    return;
-  }
-  var mid = Math.floor(((first - last) * 0.5)) + last;
-  var result = boolFunction(list[mid]);
-  if (result < 0) {
-    return search(list, boolFunction, first, mid - 1);
-  } else if (result > 0) {
-    return search(list, boolFunction, mid + 1, last);
-  } else {
-    return list[mid];
   }
 }
